+++ client/app/containers/EstablishClaimPage/EstablishClaimContainer.jsx
@@ -10,8 +10,8 @@ import { LOGO_COLORS } from '../../constants/AppConstants';
 import EstablishClaim from './EstablishClaim';
 import EstablishClaimComplete from './EstablishClaimComplete';
 import EstablishClaimCancel from './EstablishClaimCanceled';
-import UnpreparedTasksIndex from '../UnpreparedTasksIndex';
-import CanceledTasksIndex from '../CanceledTasksIndex';
+import UnpreparedTasksIndex from './UnpreparedTasksIndex';
+import CanceledTasksIndex from './CanceledTasksIndex';
 
 import TestPage from '../TestPage';
