+++ client/app/containers/EstablishClaimPage/UnpreparedTasksIndex.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Table from '../components/Table';
-import { formatDate } from '../util/DateUtil';
+import Table from '../../components/Table';
+import { formatDate } from '../../util/DateUtil';
 
 export default class UnpreparedTasksIndex extends React.Component {
   render() {
